% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goNamespace.R
\name{goNamespace}
\alias{goNamespace}
\title{Obtain a GO term's namespace}
\usage{
goNamespace(goid)
}
\arguments{
\item{goid}{A vector or GO identifiers.}
}
\value{
A named \code{character} vector of length equal to
    \code{length(goid)} containing the namespaces of the
    \code{goid} identifiers. A the identifier can't be found,
    \code{NA} is returned.
}
\description{
This function retrieves the namespaces of GO identifiers. If one
can't be found, it returns \code{NA} instead.
}
\examples{
goNamespace(c("GO:2001317", "GO:0000827", "GO:0000782"))
}
\author{
Laurent Gatto
}

