% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchGO.R
\name{searchGO}
\alias{searchGO}
\title{Search GO terms for a matching pattern}
\usage{
searchGO(pattern, namespace = c("CC", "BP", "MF"), ignore.case = TRUE,
  interactive = TRUE, graphics = getOption("menu.graphics"), ...)
}
\arguments{
\item{pattern}{A \code{character} of length 1 (subsequent elements
are ignored with a warning) to search for in GO.}

\item{namespace}{One or multiple GO namespaces to search for
\code{pattern}. Must match \code{"CC"}, \code{"MF"} and
\code{"BP"}.}

\item{ignore.case}{Should case be ignored. Default is \code{TRUE}.}

\item{interactive}{If \code{TRUE} (default) and several GO terms
matched \code{pattern}, a menu is provided to the user to
select one term.}

\item{graphics}{Should a graphical menu be used. Defaults to
\code{getOption("menu.graphics")}.}

\item{...}{Additional parameters, except \code{ignore.case} (see
above), passed to \code{\link{grep}}.}
}
\value{
A \code{character} with the identifiers of matched GO
    terms of invisibly \code{NULL} if no matches were found.
}
\description{
This function searches the the Gene Ontology terms for a specific
pattern. If run in interactive mode and if the pattern matches
more then one term, the user is presented in a menu to choose one
term.
}
\details{
This function searches the identifier, term and synonym fields of
all Gene Ontology or only a subset of the possible namespaces. The
pattern is cases insensitive, unless otherwise stated, and more
pattern matching can be further parametrised by passing additional
arguments to the underlying \code{\link{grep}} function.

If more than one match is found and the function is used
interactively, the user is presented with a menu to select a
single term. If the users exits the menu without any selection (by
choosing 0 or cancel), all matching identifiers are returned, as
if the search was performed in non-interactive mode.

If not match is found, \code{NULL} is returned invisibly with a
message.
}
\examples{
i <- interactive()
searchGO("Golgi", interactive = i)
searchGO("Golgi apparatus", interactive = i)
searchGO("Golgi apparatus", namespace = "CC", interactive = i)
searchGO("Golgi apparatus", namespace = "BP", interactive = i)
searchGO("Golgi apparatus", namespace = "MF", interactive = i)
}
\author{
Laurent Gatto
}

