% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validGO.R
\name{validGO}
\alias{validGO}
\title{Checks the validity of GO identifiers}
\usage{
validGO(goid, namespace)
}
\arguments{
\item{goid}{A vector of GO identifier.}

\item{namespace}{One of several of \code{"CC"}, \code{"MF"} or
\code{"BP"}.}
}
\value{
\code{TRUE} if \code{goid} is a valid GO identifier,
    \code{FALSE} otherwise.
}
\description{
This function checks if a GO identifier is valid by looking it up
in the \code{GO.db} package's \code{\link{GOTERM}} term
map. Depending whether a namespace is provided, the function will
look in one, multiple or all namespaces.
}
\examples{
## GO:0000014 is a valid MF Term
validGO("GO:0000014")
validGO("GO:0000014", "CC")
validGO("GO:0000014", "MF")
validGO("GO:0000014", c("CC", "MF"))
## GO:000000 is not valid
validGO("GO:0000000")
## GO:0000015 is a valid CC Term
validGO(c("GO:0000014", "GO:0000015"))
validGO(c("GO:0000014", "GO:0000015"), "CC")
validGO(c("GO:0000014", "GO:0000015"), "MF")
validGO(c("GO:0000014", "GO:0000015"), c("CC", "MF"))
validGO(c("GO:0000014", "GO:0000000"))
}
\author{
Laurent Gatto
}

